split silvan  level, swap and reordering into separate files
