+++ src/sylvan.h
@@ -59,7 +59,7 @@ namespace sylvan {
 #include <sylvan_mtbdd.h>
 #include <sylvan_bdd.h>
 #include <sylvan_ldd.h>
-#include <sylvan_reorder.h>
+#include <sylvan_var_reorder.h>
 #include <sylvan_zdd.h>
 
 #ifdef __cplusplus
