+++ test/CMakeLists.txt
@@ -16,4 +16,4 @@ add_test_file(common common.c)
 add_test_file(test_basic test_basic.c)
 add_test_file(test_cxx test_cxx.cpp)
 add_test_file(test_zdd test_zdd.c)
-add_test_file(test_reorder test_reorder.c)
+add_test_file(test_reorder test_var_swap.c)
